requirements.txt
fastapi==0.115.5
python-multipart==0.0.19
pydantic==2.10.2
uvicorn==0.32.1
openai-whisper==20240930